IteratedProcessSimulations

Documentation for IteratedProcessSimulations.

IteratedProcessSimulations.IteratedProcessSimulation

IteratedProcessSimulations.generate_data

IteratedProcessSimulations.run_simulation

IteratedProcessSimulations.run_simulation

IteratedProcessSimulations.validate_simulation_description

IteratedProcessSimulations.IteratedProcessSimulation — Type

IteratedProcessSimulation(data_generating_process, simulation_description, fit_model, summarize_model, choose_observations)

A type which collects all elements necessary to run an iterated simulation of a machine learning process.

The following attributes must be supplied:

IteratedProcessSimulations.generate_data — Method

generate_data(data_generating_process, epoch_parameters)

Generate data from a datageneratingprocess for a single epoch.

IteratedProcessSimulations.run_simulation — Method

run_simulation

Run an iterated process simulation n_simulations times.

IteratedProcessSimulations.run_simulation — Method

run_simulation(ips)

Run a single iterated process simulation.

IteratedProcessSimulations.validate_simulation_description — Method

Check whether simulationdescription dataframe is properly specified with ndatapoints column for i.i.d. sampling from data generating process