IteratedProcessSimulations

Documentation for IteratedProcessSimulations.

IteratedProcessSimulations.IteratedProcessSimulation

IteratedProcessSimulations.generate_data

IteratedProcessSimulations.generate_data

IteratedProcessSimulations.run_simulation

IteratedProcessSimulations.run_simulation

IteratedProcessSimulations.validate_inputs

IteratedProcessSimulations.validate_simulation_description

IteratedProcessSimulations.IteratedProcessSimulation — Type

IteratedProcessSimulation(data_generating_process, simulation_description, fit_model, summarize_model, choose_observations)

A type which collects all elements necessary to run an iterated simulation of a machine learning process.

The following attributes must be supplied:

IteratedProcessSimulations.generate_data — Method

generate_data(data_generating_process::Soss.Model, epoch_parameters::DataFrameRow)

Generate data from a data_generating_process for a single epoch.

IteratedProcessSimulations.generate_data — Method

generate_data(data_generating_process::Soss.Model, epoch_parameters::DataFrame)

Generate data from a data_generating_process for a series of epochs.

IteratedProcessSimulations.run_simulation — Method

run_simulation(ips::IteratedProcessSimulation, n_simulations::Int)

Run an iterated process simulation n_simulations times.

IteratedProcessSimulations.run_simulation — Method

run_simulation(ips::IteratedProcessSimulation)

Run a single iterated process simulation.

IteratedProcessSimulations.validate_inputs — Method

validate_inputs(ips::IteratedProcessSimulation)

Validate IteratedProcessSimulation object.

IteratedProcessSimulations.validate_simulation_description — Method

Check whether simulation_description dataframe is properly specified with n_datapoints column for i.i.d. sampling from data generating process